requirements.txt
flask==2.0.1
google-generativeai==0.3.0
scikit-learn==1.0.2
numpy==1.21.2 